+++ src/GRA.Controllers/ViewModel/MissionControl/Drawing/CriterionListViewModel.cs
@@ -7,5 +7,8 @@ namespace GRA.Controllers.ViewModel.MissionControl.Drawing
     {
         public IEnumerable<GRA.Domain.Model.DrawingCriterion> Criteria { get; set; }
         public PaginateViewModel PaginateModel { get; set; }
+
+        public string Search { get; set; }
+        public bool? Mine { get; set; }
     }
 }