+++ src/GRA.Data/Repository/DrawingCriterionRepository.cs
@@ -1,12 +1,13 @@
 ﻿using AutoMapper.QueryableExtensions;
 using GRA.Domain.Model;
+using GRA.Domain.Model.Filters;
 using GRA.Domain.Repository;
+using GRA.Domain.Repository.Extensions;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using System;
 
 namespace GRA.Data.Repository
 {
@@ -19,26 +20,41 @@ namespace GRA.Data.Repository
         {
         }
 
-        public async Task<IEnumerable<DrawingCriterion>> PageAllAsync(int siteId, int skip, int take)
+        public async Task<IEnumerable<DrawingCriterion>> PageAllAsync(BaseFilter filter)
         {
-            return await DbSet
-                    .AsNoTracking()
-                    .Include(_ => _.Branch)
-                    .Where(_ => _.SiteId == siteId)
-                    .OrderBy(_ => _.Name)
-                    .ThenBy(_ => _.Id)
-                    .Skip(skip)
-                    .Take(take)
-                    .ProjectTo<DrawingCriterion>()
-                    .ToListAsync();
+            return await ApplyFilters(filter)
+                .OrderBy(_ => _.Name)
+                .ThenBy(_ => _.Id)
+                .ApplyPagination(filter)
+                .ProjectTo<DrawingCriterion>()
+                .ToListAsync();
         }
 
-        public async Task<int> GetCountAsync(int siteId)
+        public async Task<int> GetCountAsync(BaseFilter filter)
         {
-            return await DbSet
-                .AsNoTracking()
-                .Where(_ => _.SiteId == siteId)
+            return await ApplyFilters(filter)
                 .CountAsync();
+        }
+
+        private IQueryable<Model.DrawingCriterion> ApplyFilters(BaseFilter filter)
+        {
+            var criterionList = DbSet
+                .AsNoTracking()
+                .Where(_ => _.SiteId == filter.SiteId);
+
+            if (filter.UserIds?.Any() == true)
+            {
+                criterionList = criterionList.Where(_ => filter.UserIds.Contains(_.CreatedBy));
+            }
+
+            if (!string.IsNullOrWhiteSpace(filter.Search))
+            {
+                criterionList = criterionList.Where(_ => _.Name.Contains(filter.Search))
+                    .Union(criterionList.Where(_ => _.System.Name.Contains(filter.Search)))
+                    .Union(criterionList.Where(_ => _.Branch.Name.Contains(filter.Search)));
+            }
+
+            return criterionList;
         }
 
         public async Task<int> GetEligibleUserCountAsync(int criterionId)
@@ -87,7 +103,7 @@ namespace GRA.Data.Repository
                     .Select(_ => _.UserId);
 
                 users = users.Where(_ => !previousWinners.Contains(_.Id));
-            }           
+            }
 
             var userIds = users.Select(_ => _.Id);
             IQueryable<int> activityUsers = null;