+++ src/GRA.Domain.Model/DrawingCriterion.cs
@@ -31,6 +31,7 @@ namespace GRA.Domain.Model
         [DisplayName("Exclude Previous Winners")]
         public bool ExcludePreviousWinners { get; set; }
 
+        public string SystemName { get; set; }
         public string BranchName { get; set; }
         [DisplayName("Eligible Count")]
         public int EligibleCount { get; set; }